Extra all features from content.js
