+++ src/features/add-confirmation-to-comment-cancellation.js
@@ -1,0 +1,15 @@
+// Prompt user to confirm erasing a comment with the Cancel button
+export default function () {
+	$(document).on('click', '.js-hide-inline-comment-form', event => {
+		// Do not prompt if textarea is empty
+		const textarea = event.target.closest('.js-inline-comment-form').querySelector('.js-comment-field');
+		if (textarea.value === '') {
+			return;
+		}
+
+		if (window.confirm('Are you sure you want to discard your unsaved changes?') === false) { // eslint-disable-line no-alert
+			event.stopPropagation();
+			event.stopImmediatePropagation();
+		}
+	});
+}
