+++ src/features/add-milestone-navigation.js
@@ -1,0 +1,17 @@
+import {h} from 'dom-chef';
+import select from 'select-dom';
+import * as pageDetect from '../libs/page-detect';
+
+const repoUrl = pageDetect.getRepoURL();
+
+export default function () {
+	select('.repository-content').before(
+		<div class="subnav">
+			<div class="subnav-links float-left" role="navigation">
+				<a href={`/${repoUrl}/labels`} class="subnav-item">Labels</a>
+				<a href={`/${repoUrl}/milestones`} class="subnav-item">Milestones</a>
+			</div>
+		</div>
+	);
+}
+
