+++ src/features/add-patch-diff-links.js
@@ -1,0 +1,24 @@
+import {h} from 'dom-chef';
+import select from 'select-dom';
+import * as pageDetect from '../libs/page-detect';
+
+export default function () {
+	if (select.exists('.sha-block.patch-diff-links')) {
+		return;
+	}
+
+	let commitUrl = location.pathname.replace(/\/$/, '');
+
+	if (pageDetect.isPRCommit()) {
+		commitUrl = commitUrl.replace(/\/pull\/\d+\/commits/, '/commit');
+	}
+
+	select('.commit-meta span.float-right').append(
+		<span class="sha-block patch-diff-links">
+			<a href={`${commitUrl}.patch`} class="sha">patch</a>
+			{ ' ' /* Workaround for: JSX eats whitespace between elements */ }
+			<a href={`${commitUrl}.diff`} class="sha">diff</a>
+		</span>
+	);
+}
+
