+++ src/features/add-title-to-emojis.js
@@ -1,0 +1,8 @@
+import select from 'select-dom';
+
+export default function () {
+	for (const emoji of select.all('g-emoji')) {
+		emoji.setAttribute('title', `:${emoji.getAttribute('alias')}:`);
+	}
+}
+
