+++ src/features/hide-empty-meta.js
@@ -1,0 +1,12 @@
+import select from 'select-dom';
+import * as pageDetect from '../libs/page-detect';
+
+export default function () {
+	if (pageDetect.isRepoRoot()) {
+		const meta = select('.repository-meta');
+		if (select.exists('em', meta) && !select.exists('.js-edit-repo-meta-button')) {
+			meta.style.display = 'none';
+		}
+	}
+}
+
