+++ src/features/open-ci-details-in-new-tab.js
@@ -1,0 +1,10 @@
+import select from 'select-dom';
+
+export default function () {
+	const CIDetailsLinks = select.all('a.status-actions');
+	for (const link of CIDetailsLinks) {
+		link.setAttribute('target', '_blank');
+		link.setAttribute('rel', 'noopener');
+	}
+}
+
