+++ src/features/remove-projects-tab.js
@@ -1,0 +1,9 @@
+import select from 'select-dom';
+
+export default function () {
+	const projectsTab = select('.js-repo-nav .reponav-item[data-selected-links^="repo_projects"]');
+	if (projectsTab && projectsTab.querySelector('.Counter, .counter').textContent === '0') {
+		projectsTab.remove();
+	}
+}
+
